rand: use size_t for size argument to RAND_bytes_ex()
(Merged from https://github.com/openssl/openssl/pull/15540)

+++ include/openssl/rand.h
@@ -65,14 +65,14 @@ int RAND_priv_bytes(unsigned char *buf, int num);
  * Equivalent of RAND_priv_bytes() but additionally taking an OSSL_LIB_CTX and
  * a strength.
  */
-int RAND_priv_bytes_ex(OSSL_LIB_CTX *ctx, unsigned char *buf, int num,
+int RAND_priv_bytes_ex(OSSL_LIB_CTX *ctx, unsigned char *buf, size_t num,
                        unsigned int strength);
 
 /*
  * Equivalent of RAND_bytes() but additionally taking an OSSL_LIB_CTX and
  * a strength.
  */
-int RAND_bytes_ex(OSSL_LIB_CTX *ctx, unsigned char *buf, int num,
+int RAND_bytes_ex(OSSL_LIB_CTX *ctx, unsigned char *buf, size_t num,
                   unsigned int strength);
 
 # ifndef OPENSSL_NO_DEPRECATED_1_1_0
